Chore - Refactor utils files (#4220)
* Chore - refactor util files for separation of concerns
